(re)refactor(adapter): dataset spec for each worker
me di cuenta de que es mejor que el tema de la división del dataset se
haga a nivel config y que después adaptar dataset config a dataset spec
sí sea uno a uno. Después lo cambio

+++ comms/src/share_dataset/recv_dataset.rs
@@ -19,7 +19,7 @@ where
 
     let mut received = 0;
 
-    while received < spec.size {
+    while (received as u64) < spec.size {
         let msg: Msg = receiver.recv_into(&mut buf).await?;
 
         let chunk = match msg {
